requirements.txt
pandas>=2.1,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.4,<2.0
torch>=2.2,<3.0
sentence-transformers>=3.0,<4.0
transformers>=4.40,<5.0
pyyaml>=6.0,<7.0
jsonlines>=4.0,<5.0
tqdm>=4.66,<5.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0
pytest>=8.0,<9.0
